refactor: only load internal Spark component as internal
