+++ sdk/ml/azure-ai-ml/azure/ai/ml/_internal/entities/node.py
@@ -295,99 +295,3 @@ class Sweep(InternalBaseNode):
     def __init__(self, **kwargs):
         kwargs.pop("type", None)
         super(Sweep, self).__init__(type=NodeType.SWEEP, **kwargs)
-
-
-class Spark(InternalBaseNode):
-    def __init__(self, **kwargs):
-        kwargs.pop("type", None)
-        super(Spark, self).__init__(type=NodeType.SPARK, **kwargs)
-        self._init = True
-        self._compute_name: str = kwargs.pop("compute_name", None)
-        self._queue: str = kwargs.pop("queue", None)
-        self._driver_memory: str = kwargs.pop("driver_memory", None)
-        self._driver_cores: int = kwargs.pop("driver_cores", None)
-        self._executor_memory: str = kwargs.pop("executor_memory", None)
-        self._executor_cores: int = kwargs.pop("executor_cores", None)
-        self._number_executors: int = kwargs.pop("number_executors", None)
-        self._conf: Union[dict, str] = kwargs.pop("conf", None)
-        self._init = False
-
-    @property
-    def compute_name(self) -> str:
-        """Name of the compute to be used."""
-        return self._compute_name
-
-    @compute_name.setter
-    def compute_name(self, value: str):
-        self._compute_name = value
-
-    @property
-    def queue(self) -> str:
-        """The name of the YARN queue to which submitted."""
-        return self._queue
-
-    @queue.setter
-    def queue(self, value: str):
-        self._queue = value
-
-    @property
-    def driver_memory(self) -> str:
-        """Amount of memory to use for the driver process.
-
-        It's the same format as JVM memory strings. Use lower-case suffixes, e.g. k, m, g, t, and p, for kilobyte,
-        megabyte, gigabyte and terabyte respectively. Example values are 10k, 10m and 10g.
-        """
-        return self._driver_memory
-
-    @driver_memory.setter
-    def driver_memory(self, value: str):
-        self._driver_memory = value
-
-    @property
-    def driver_cores(self) -> int:
-        """Number of cores to use for the driver process."""
-        return self._driver_cores
-
-    @driver_cores.setter
-    def driver_cores(self, value: int):
-        self._driver_cores = value
-
-    @property
-    def executor_memory(self) -> str:
-        """Amount of memory to use per executor process.
-
-        It's the same format as JVM memory strings. Use lower-case suffixes, e.g. k, m, g, t, and p, for kilobyte,
-        megabyte, gigabyte and terabyte respectively. Example values are 10k, 10m and 10g.
-        """
-        return self._executor_memory
-
-    @executor_memory.setter
-    def executor_memory(self, value: str):
-        self._executor_memory = value
-
-    @property
-    def executor_cores(self) -> int:
-        """Number of cores to use for each executor."""
-        return self._executor_cores
-
-    @executor_cores.setter
-    def executor_cores(self, value: int):
-        self._executor_cores = value
-
-    @property
-    def number_executors(self) -> int:
-        """Number of executors to launch for this session."""
-        return self._number_executors
-
-    @number_executors.setter
-    def number_executors(self, value: int):
-        self._number_executors = value
-
-    @property
-    def conf(self) -> Union[dict, str]:
-        """Spark configuration properties."""
-        return self._conf
-
-    @conf.setter
-    def conf(self, value: Union[dict, str]):
-        self._conf = value
