+++ sdk/ml/azure-ai-ml/azure/ai/ml/entities/_component/component_factory.py
@@ -6,24 +6,28 @@
 
 from typing import Any, Callable, Dict, Optional, Tuple
 
-from marshmallow import Schema
-
 from azure.ai.ml._restclient.v2022_05_01.models import ComponentVersionData
-from azure.ai.ml.constants._common import SOURCE_PATH_CONTEXT_KEY
-from azure.ai.ml.constants._component import NodeType, DataTransferTaskType
+from azure.ai.ml._utils.utils import is_internal_components_enabled
+from azure.ai.ml.constants._common import (
+    AZUREML_INTERNAL_COMPONENTS_SCHEMA_PREFIX,
+    SOURCE_PATH_CONTEXT_KEY,
+    CommonYamlFields,
+)
+from azure.ai.ml.constants._component import DataTransferTaskType, NodeType
 from azure.ai.ml.entities._component.automl_component import AutoMLComponent
 from azure.ai.ml.entities._component.command_component import CommandComponent
 from azure.ai.ml.entities._component.component import Component
+from azure.ai.ml.entities._component.datatransfer_component import (
+    DataTransferCopyComponent,
+    DataTransferExportComponent,
+    DataTransferImportComponent,
+)
 from azure.ai.ml.entities._component.import_component import ImportComponent
 from azure.ai.ml.entities._component.parallel_component import ParallelComponent
 from azure.ai.ml.entities._component.pipeline_component import PipelineComponent
 from azure.ai.ml.entities._component.spark_component import SparkComponent
-from azure.ai.ml.entities._component.datatransfer_component import (
-    DataTransferCopyComponent,
-    DataTransferImportComponent,
-    DataTransferExportComponent,
-)
 from azure.ai.ml.entities._util import get_type_from_spec
+from marshmallow import Schema
 
 
 class _ComponentFactory:
@@ -81,10 +85,23 @@ class _ComponentFactory:
             create_schema_func=DataTransferExportComponent._create_schema_for_validation,
         )
 
-    def get_create_funcs(self, yaml_spec: dict) -> Tuple[Callable[..., Component], Callable[[Any], Schema]]:
+    def get_create_funcs(
+        self, yaml_spec: dict, for_load=False
+    ) -> Tuple[Callable[..., Component], Callable[[Any], Schema]]:
         """Get registered functions to create instance & its corresponding schema for the given type."""
 
         _type = get_type_from_spec(yaml_spec, valid_keys=self._create_instance_funcs)
+        if for_load and is_internal_components_enabled():
+            schema_url = yaml_spec[CommonYamlFields.SCHEMA] if CommonYamlFields.SCHEMA in yaml_spec else None
+            if (
+                _type == NodeType.SPARK
+                and schema_url
+                and schema_url.startswith(AZUREML_INTERNAL_COMPONENTS_SCHEMA_PREFIX)
+            ):
+                from azure.ai.ml._internal._schema.node import NodeType as InternalNodeType
+
+                _type = InternalNodeType.INTERNAL_SPARK
+
         create_instance_func = self._create_instance_funcs[_type]
         create_schema_func = self._create_schema_funcs[_type]
         return create_instance_func, create_schema_func
